+++ Core/Account/DTOS/AccountDto.cs
@@ -3,7 +3,7 @@ namespace Core.Account.DTOS;
 public class AccountDto
 {
     public Guid Id { get; set; }
-    public decimal Pesel { get; set; }
+    public long Pesel { get; set; }
     public string Name { get; set; }
     public string? Secont_name { get; set; }
     public string Last_name { get; set; }
